chatuba

+++ routes/teams.js
@@ -28,12 +28,26 @@ router.post('/new',(req,res) =>{
       coach: req.body.coach,
       capitan: req.body.capitan,
     }
-    models.Team.create(teamData).then(team => {
+    var decoded = jwt.verify(req.headers['x-access-token'], process.env.SECRET_KEY)
+  models.User.findOne({
+    where: {
+      id: decoded.id
+    }
+    }).then(user =>{
+        if(user.isAdmin){
+            models.Team.create(teamData).then(team => {
                 res.json({ status: team.name + ' Registered!'})
               })
               .catch(err => {
                 res.send('error: ' + err)
               })
+        }else{
+            res.send('tu não és admin parceiro')
+        }
+    })
+    .catch(err =>{
+        res.send('error: ' + err)
+    })
   });
 
 
@@ -53,14 +67,27 @@ router.post('/new',(req,res) =>{
 
 //delete team by id
 router.delete('/del/',(req,res)=> {
-    console.log(req.body.id)
-    models.Team.destroy({
-        where: {
-            id: req.body.id
+    var decoded = jwt.verify(req.headers['x-access-token'], process.env.SECRET_KEY)
+  models.User.findOne({
+    where: {
+      id: decoded.id
+    }
+    }).then(user => {
+        if (user.isAdmin){
+
+            console.log(req.body.id)
+            models.Team.destroy({
+                where: {
+                    id: req.body.id
+                }
+            })
+            .then(resp => 
+              {res.json(resp)})
+    
         }
-    })
-    .then(resp => 
-      {res.json(resp)})
+        }).catch(err => {
+            res.send(err)
+        })
 
 })
 
@@ -68,32 +95,45 @@ router.delete('/del/',(req,res)=> {
 //update team data by id
 router.put('/update/',(req,res) =>
 {
-  models.Team.findOne({
-    where: {
-      id: req.body.id
-    }
-    })
-    .then(team =>{
-    const teamData = {
-        name: req.body.name ? req.body.name : team.name,
-      coach: req.body.coach ? req.body.coach : team.coach,
-      capitan: req.body.capitan ? req.body.capitan : team.capitan
-    }
-    console.log('até aqui foi')
-    models.Team.update(
-      {name: teamData.name,
-        coach: teamData.coach,
-        capitan: teamData.capitan
-      },
-      {where: {id : team.id}}
-    ).then(newTeam => {res.send(newTeam)})
-    .catch(res.send(team))
-
-
-    })
-    .catch(res => {
-      return res
+    var decoded = jwt.verify(req.headers['x-access-token'], process.env.SECRET_KEY)
+    models.User.findOne({
+      where: {
+        id: decoded.id
+      }
+      }).then(user => {
+        if(user.isAdmin){
+            models.Team.findOne({
+                where: {
+                  id: req.body.id
+                }
+                })
+                .then(team =>{
+                const teamData = {
+                    name: req.body.name ? req.body.name : team.name,
+                  coach: req.body.coach ? req.body.coach : team.coach,
+                  capitan: req.body.capitan ? req.body.capitan : team.capitan
+                }
+                console.log('até aqui foi')
+                models.Team.update(
+                  {name: teamData.name,
+                    coach: teamData.coach,
+                    capitan: teamData.capitan
+                  },
+                  {where: {id : team.id}}
+                ).then(newTeam => {res.send(newTeam)})
+                .catch(res.send(team))
+            
+            
+                })
+                .catch(res => {
+                  return res
+                })
+        }
+      })
+      .catch(err => {
+        res.send(err)
     })
   });
 
 module.exports = router;
+
